+++ chalk-ir/src/fold.rs
@@ -4,6 +4,7 @@ use crate::family::TargetTypeFamily;
 use crate::*;
 use std::fmt::Debug;
 
+mod binder_impls;
 mod boring_impls;
 pub mod shift;
 mod subst;
@@ -387,70 +388,5 @@ impl<TF: TypeFamily, TTF: TargetTypeFamily<TF>> Fold<TF, TTF> for Lifetime<TF> {
         binders: usize,
     ) -> Fallible<Self::Result> {
         folder.fold_lifetime(self, binders)
-    }
-}
-
-impl<TF: TypeFamily, TTF: TargetTypeFamily<TF>> Fold<TF, TTF> for QuantifiedTy<TF> {
-    type Result = QuantifiedTy<TTF>;
-    fn fold_with(
-        &self,
-        folder: &mut dyn Folder<TF, TTF>,
-        binders: usize,
-    ) -> Fallible<Self::Result> {
-        let QuantifiedTy {
-            num_binders,
-            ref ty,
-        } = *self;
-        Ok(QuantifiedTy {
-            num_binders,
-            ty: ty.fold_with(folder, binders + num_binders)?,
-        })
-    }
-}
-
-impl<T, TF: TypeFamily, TTF: TargetTypeFamily<TF>> Fold<TF, TTF> for Binders<T>
-where
-    T: Fold<TF, TTF>,
-    TF: TypeFamily,
-{
-    type Result = Binders<T::Result>;
-    fn fold_with(
-        &self,
-        folder: &mut dyn Folder<TF, TTF>,
-        binders: usize,
-    ) -> Fallible<Self::Result> {
-        let Binders {
-            binders: ref self_binders,
-            value: ref self_value,
-        } = *self;
-        let value = self_value.fold_with(folder, binders + self_binders.len())?;
-        Ok(Binders {
-            binders: self_binders.clone(),
-            value: value,
-        })
-    }
-}
-
-impl<T, TF, TTF> Fold<TF, TTF> for Canonical<T>
-where
-    T: Fold<TF, TTF>,
-    TF: TypeFamily,
-    TTF: TargetTypeFamily<TF>,
-{
-    type Result = Canonical<T::Result>;
-    fn fold_with(
-        &self,
-        folder: &mut dyn Folder<TF, TTF>,
-        binders: usize,
-    ) -> Fallible<Self::Result> {
-        let Canonical {
-            binders: ref self_binders,
-            value: ref self_value,
-        } = *self;
-        let value = self_value.fold_with(folder, binders + self_binders.len())?;
-        Ok(Canonical {
-            binders: self_binders.clone(),
-            value: value,
-        })
     }
 }
